Factor out the server
The server component got quite large. Factor it out into a separate
module. Wrap it in a class PeekabooServer for additional containment.
